+++ unittest/google_test.cpp
@@ -9,9 +9,9 @@
 #include "cppknife_control.hpp"
 
 int googleTest(int argc, char **argv) {
-#ifdef USE_UNITTESTS
-  testing::InitGoogleTest (&argc, argv);
-  return RUN_ALL_TESTS ();
+#ifndef NOT_USE_UNITTESTS
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
 #else
   fprintf(stderr, "No unittest configured!\n");
   return 0;